requirements.txt
streamlit==1.28.1
pandas>=2.2.0
numpy>=1.26.0
pillow>=10.2.0
jinja2>=3.1.2
qrcode[pil]>=7.4.2
aiofiles>=23.2.1
sqlalchemy>=2.0.23
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.31.0
pydantic>=2.6.0
plotly>=5.18.0
streamlit-option-menu>=0.3.6
